Remove --Xfork-choice-before-proposing-enabled feature toggle.
It's been enabled by default everywhere for a while now and is working well.
