--- ethereum/statetransition/src/main/java/tech/pegasys/teku/statetransition/forkchoice/PreProposalForkChoiceTrigger.java
@@ -1,32 +1,0 @@
-/*
- * Copyright ConsenSys Software Inc., 2022
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-package tech.pegasys.teku.statetransition.forkchoice;
-
-import tech.pegasys.teku.infrastructure.async.SafeFuture;
-import tech.pegasys.teku.infrastructure.unsigned.UInt64;
-
-public class PreProposalForkChoiceTrigger extends ForkChoiceTrigger {
-
-  private final ForkChoice forkChoice;
-
-  public PreProposalForkChoiceTrigger(final ForkChoice forkChoice) {
-    super(forkChoice);
-    this.forkChoice = forkChoice;
-  }
-
-  @Override
-  public SafeFuture<Void> prepareForBlockProduction(final UInt64 slot) {
-    return forkChoice.prepareForBlockProduction(slot);
-  }
-}
