+++ ethereum/statetransition/src/test/java/tech/pegasys/teku/statetransition/forkchoice/ForkChoiceTriggerTest.java
@@ -13,12 +13,11 @@
 
 package tech.pegasys.teku.statetransition.forkchoice;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
+import static tech.pegasys.teku.infrastructure.async.SafeFutureAssert.assertThatSafeFuture;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -48,8 +47,14 @@ class ForkChoiceTriggerTest {
   }
 
   @Test
-  void shouldNotRunForkChoicePriorToBlockProduction() {
-    assertThat(trigger.prepareForBlockProduction(UInt64.ONE)).isCompleted();
-    verifyNoInteractions(forkChoice);
+  void shouldRunForkChoicePriorToBlockProduction() {
+    final SafeFuture<Void> prepareFuture = new SafeFuture<>();
+    when(forkChoice.prepareForBlockProduction(any())).thenReturn(prepareFuture);
+    final SafeFuture<Void> result = trigger.prepareForBlockProduction(UInt64.ONE);
+    assertThatSafeFuture(result).isNotDone();
+    verify(forkChoice).prepareForBlockProduction(UInt64.ONE);
+
+    prepareFuture.complete(null);
+    assertThatSafeFuture(result).isCompleted();
   }
 }
