+++ teku/src/test/java/tech/pegasys/teku/cli/options/Eth2NetworkOptionsTest.java
@@ -27,7 +27,6 @@ import org.junit.jupiter.params.provider.ValueSource;
 import tech.pegasys.teku.cli.AbstractBeaconNodeCommandTest;
 import tech.pegasys.teku.config.TekuConfiguration;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
-import tech.pegasys.teku.networks.Eth2NetworkConfiguration;
 import tech.pegasys.teku.spec.Spec;
 import tech.pegasys.teku.spec.SpecMilestone;
 import tech.pegasys.teku.spec.config.ProgressiveBalancesMode;
@@ -205,22 +204,6 @@ class Eth2NetworkOptionsTest extends AbstractBeaconNodeCommandTest {
         () ->
             getTekuConfigurationFromArguments(
                 "--Xnetwork-terminal-block-hash-epoch-override", "asd"));
-  }
-
-  @Test
-  void shouldEnableForkChoiceBeforeProposing() {
-    final Eth2NetworkConfiguration networkConfig =
-        getTekuConfigurationFromArguments("--Xfork-choice-before-proposing-enabled")
-            .eth2NetworkConfiguration();
-    assertThat(networkConfig.isForkChoiceBeforeProposingEnabled()).isTrue();
-  }
-
-  @Test
-  void shouldDisableForkChoiceBeforeProposing() {
-    final Eth2NetworkConfiguration networkConfig =
-        getTekuConfigurationFromArguments("--Xfork-choice-before-proposing-enabled=false")
-            .eth2NetworkConfiguration();
-    assertThat(networkConfig.isForkChoiceBeforeProposingEnabled()).isFalse();
   }
 
   @Test
